setting refugees tooltip items
