+++ src/components/PopUp/PopUpProject.js
@@ -24,25 +24,8 @@ class PopUpProject extends PopUp {
             </ul>`
   }
 
-  _getRefugeesInfo() {
-    const title = !(this.model.get('totals')['projects']) ? `<h1 class="text text-module-title -light"> ${this.model.get('location')['name']} - <span class="title-sector text text-legend-s -light">${this.model.get('crisis')[0]['name']}</span></h1>
-      <hr></hr>`: `<p class="text text-report-title -light">${this.model.get('crisis')[0]['name']}</p>`
-
-    return `
-      ${title}
-      <div class="refugees-info">
-        <span class="title-sector text text-legend-s -light">Countries with projects for refugee assistance in the country</span>
-        <ul>
-          ${ this.model.get('crisis')[0] && this.model.get('crisis')[0]['parties_involved'].length > 0 ? this.model.get('crisis')[0]['parties_involved'].map(country => `<li class="sector text text-legend-s -light"> ${country.country} </li>`).join('') : ''}
-        </ul>
-      </div>
-      <hr></hr>
-    `
-  }
-
   _getProjectsInfo() {
     const sectorsItems = (this.model.get('sectors').length > 0) ? this._getSectors() : '';
-    const refugeeTitle = this.model.get('crisis')[0] ? `<p class="title-sector text text-legend-s -light">${this.model.get('crisis')[0]['name']}</p>` : '';
 
     return `
       <h1 class="text text-module-title -light"> ${this.model.get('location')['name']} - <span class="number-m">${ utils.numberNotation(  this.model.get('totals')['projects'] )} projects in ${this.model.get('year')}
@@ -59,7 +42,6 @@ class PopUpProject extends PopUp {
   }
 
   _getContent() {
-    const refuggesInfo = (this.model.get('crisis')[0]) ? this._getRefugeesInfo() : '';
     const projectsInfo = (this.model.get('totals')['projects']) ? this._getProjectsInfo() : '';
 
     return `<div class="wrapper -project">
