+++ src/scripts/models/RefugeesModel.js
@@ -1,0 +1,15 @@
+'use strict';
+
+import $ from 'jquery';
+import PopUpModel from './PopUpModel';
+import filtersModel from '../../scripts/models/filtersModel';
+
+class RefugeesModel extends PopUpModel {
+
+  constructor(options) {
+    super(options);
+  }
+}
+
+export default RefugeesModel;
+
